Add a few explicit conversions to size_t in device/fido.
base::span will shortly require that length is unsigned.

This CL fixes two minor issues in device/fido related to this
upcoming change.

1. SHA256_DIGEST_LENGTH, as defined, is unfortunately untyped
so a static cast is added in this CL to make it a size_t
correctly so it can be passed into make::span.
2. In virtual ctap2 device, GetUnsigned returns a uint64_t which
might not fit into a size_t on 32-bit platforms. Thus a
checked_cast is added to handle this.
3. The manual pointer arithmetic to obtain the response key is
replaced by a call to subspan.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220318
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101081}

+++ device/fido/pin_internal.cc
@@ -284,10 +284,11 @@ class ProtocolV2 : public ProtocolV1 {
     static_assert(kSharedKeyLength == 2 * SHA256_DIGEST_LENGTH, "");
     DCHECK_GE(*out_len, static_cast<size_t>(kSharedKeyLength));
     auto hmac_key_out =
-        base::make_span(reinterpret_cast<uint8_t*>(out), SHA256_DIGEST_LENGTH);
+        base::make_span(reinterpret_cast<uint8_t*>(out),
+                        static_cast<size_t>(SHA256_DIGEST_LENGTH));
     auto aes_key_out =
         base::make_span(reinterpret_cast<uint8_t*>(out) + SHA256_DIGEST_LENGTH,
-                        SHA256_DIGEST_LENGTH);
+                        static_cast<size_t>(SHA256_DIGEST_LENGTH));
 
     constexpr uint8_t kHMACKeyInfo[] = "CTAP2 HMAC key";
     constexpr uint8_t kAESKeyInfo[] = "CTAP2 AES key";
